Move record and enum into their own file
