+++ src/main/java/org/opentripplanner/updater/spi/TripTimesValidationMapper.java
@@ -2,18 +2,17 @@ package org.opentripplanner.updater.spi;
 
 import org.opentripplanner.transit.model.framework.FeedScopedId;
 import org.opentripplanner.transit.model.framework.Result;
-import org.opentripplanner.transit.model.timetable.TripTimes;
-import org.opentripplanner.updater.spi.UpdateError;
+import org.opentripplanner.transit.model.timetable.ValidationError;
 
 /**
- * Converts the result of a {@link TripTimes} to the model of the updater ready to be consumed
- * by the metrics.
+ * Converts a {@link ValidationError} to the model of the updater ready to be consumed
+ * by the metrics APIS and logs.
  */
 public class TripTimesValidationMapper {
 
   public static <T> Result<T, UpdateError> toResult(
     FeedScopedId tripId,
-    TripTimes.ValidationError validationError
+    ValidationError validationError
   ) {
     var type =
       switch (validationError.type()) {
